requirements-embeddings.txt
torch==2.4.1
fair-esm==2.0.0
numpy>=1.24
